menambah falkutas

+++ app/Http/Controllers/SettingIKUController.php
@@ -6,6 +6,7 @@ use App\Models\IndikatorKinerja;
 use App\Models\SettingIKU;
 use App\Models\tahun_kerja;
 use Illuminate\Http\Request;
+use RealRashid\SweetAlert\Facades\Alert;
 
 class SettingIKUController extends Controller
 {
@@ -36,26 +37,77 @@ class SettingIKUController extends Controller
 }
 
 
-    public function store(Request $request)
+public function store(Request $request)
     {
         $request->validate([
-            'ik_id' => 'required|exists:IndikatorKinerja,ik_id',
+            'ik_id' => 'required|exists:indikator_kinerja,ik_id',
             'th_id' => 'required|exists:tahun_kerja,th_id',
         ]);
 
+        $Setting = SettingIKU::where('ik_id', $request->ik_id)
+                ->where('th_id', $request->th_id)
+                ->first();
+
+            if ($Setting) {
+                return response()->json([
+                    'success' => false,
+                    'message' => 'Indikator ini sudah ada untuk tahun yang sama.',
+                ]);
+            }
+
         SettingIKU::create([
+            'id_setting' => 'IS' . md5(uniqid(rand(), true)),
             'ik_id' => $request->ik_id,
             'th_id' => $request->th_id,
+            'status' => 0,
         ]);
 
         return response()->json(['success' => true, 'message' => 'Setting IKU berhasil ditambahkan.']);
     }
 
-    public function destroy($id)
-    {
-        $setting = SettingIKU::findOrFail($id);
+    public function update(Request $request, $id_setting)
+{
+    $request->validate([
+        'ik_id' => 'required|exists:indikator_kinerja,ik_id',
+        'th_id' => 'required|exists:tahun_kerja,th_id',
+    ]);
+
+    $setting = SettingIKU::findOrFail($id_setting);
+    $existingSetting = SettingIKU::where('ik_id', $request->ik_id)
+        ->where('th_id', $request->th_id)
+        ->where('id_setting', '!=', $id_setting)
+        ->first();
+
+    if ($existingSetting) {
+        return response()->json([
+            'success' => false,
+            'message' => 'Indikator ini sudah ada untuk tahun yang sama.',
+        ]);
+    }
+
+    $setting->update([
+        'ik_id' => $request->ik_id,
+        'th_id' => $request->th_id,
+    ]);
+
+    return response()->json(['success' => true, 'message' => 'Setting IKU berhasil diperbarui.']);
+}
+
+
+public function destroy($id_setting)
+{
+    $setting = SettingIKU::find($id_setting);
+
+    if ($setting) {
         $setting->delete();
 
-        return redirect()->route('settingiku.index')->with('success', 'Setting IKU berhasil dihapus.');
+        Alert::success('Sukses', 'Data berhasil dihapus');
+        return redirect()->route('settingiku.index');
     }
+
+    Alert::error('Error', 'Terjadi kesalahan!');
+    return redirect()->route('settingiku.index');
+}
+
+
 }
